% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.RegLin.R
\name{plot.RegLin}
\alias{plot.RegLin}
\title{Método de Plotagem para RegLin}
\usage{
\method{plot}{RegLin}(x, ...)
}
\arguments{
\item{x}{Um objeto da classe RegLin.}

\item{...}{Outros argumentos (não utilizados).}
}
\description{
Gera gráficos diagnósticos para um modelo de regressão linear.
}
