% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.RegLin.R
\name{predict.RegLin}
\alias{predict.RegLin}
\title{Método de Predição para RegLin}
\usage{
\method{predict}{RegLin}(model, newdata)
}
\arguments{
\item{newdata}{Novo banco de dados}

\item{x}{Um objeto da classe RegLin.}

\item{...}{Outros argumentos (não utilizados).}
}
\description{
Gera Predições para um modelo de regressão linear.
}
