% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_reglin.R
\name{predict_reglin}
\alias{predict_reglin}
\title{Predição com base em um modelo de regressão múltipla}
\usage{
predict_reglin(model, newdata)
}
\arguments{
\item{model}{Lista contendo os coeficientes estimados pelo modelo de regressão múltipla.
A lista deve conter pelo menos o elemento \code{coefficients}, que são os coeficientes estimados.}

\item{newdata}{Um data frame contendo as novas observações para as quais as predições serão feitas.
As colunas devem corresponder às variáveis explicativas usadas no ajuste do modelo.}
}
\value{
Retorna um vetor de valores numéricos contendo as predições feitas com base em \code{newdata}.
}
\description{
Esta função realiza predições a partir de um modelo de regressão múltipla ajustado.
Os coeficientes do modelo são usados para prever novos valores com base em novos dados fornecidos.
}
\examples{
# Semente
set.seed(1)
# Ajuste de um modelo de regressão múltipla
model = RegLin(y ~ x1 + x2, data = exampleData)

# Novos dados para predição
newdata = data.frame(x1 = runif(2), x2 = runif(2, 10, 20))
predict_reglin(model, newdata)

}
