% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.RegLin.R
\name{print.RegLin}
\alias{print.RegLin}
\title{Método de Impressão para RegLin}
\usage{
\method{print}{RegLin}(x, ...)
}
\arguments{
\item{x}{Um objeto da classe RegLin.}

\item{...}{Outros argumentos (não utilizados).}
}
\description{
Exibe os coeficientes do modelo de forma legível.
}
