% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegLin.R
\name{RegLin}
\alias{RegLin}
\title{Ajuste de um modelo de regressão múltipla}
\usage{
RegLin(formula, data)
}
\arguments{
\item{formula}{Uma fórmula R especificando o modelo de regressão no formato \code{Y ~ X1 + X2 + ... + Xn}, onde \code{Y} é a variável dependente e \code{X1}, \code{X2}, ..., \code{Xn} são as variáveis independentes.}

\item{data}{Um data frame contendo os dados que serão utilizados para ajustar o modelo. As colunas devem corresponder às variáveis especificadas na fórmula.}
}
\value{
Retorna uma lista com os seguintes elementos:
\describe{
\item{data}{Data frame utilizado.}
\item{coefficients}{Os coeficientes estimados.}
\item{residuals}{Os resíduos do modelo.}
\item{fitted.values}{Os valores ajustados pelo modelo.}
\item{call}{Função utilizada para realizar a regressão.}
\item{formula}{Formula utilizada para realizar a regressão.}
\item{X}{A matriz de design utilizada no ajuste do modelo.}
\item{Y}{O vetor da variável dependente.}
}
}
\description{
Esta função ajusta um modelo de regressão múltipla com base na fórmula fornecida e nos dados.
Ela retorna os coeficientes estimados, os resíduos, os valores ajustados, e as variáveis dependente e independentes.
}
\examples{
# Ajuste de um modelo de regressão múltipla
model = RegLin(y ~ x1 + x2, data = exampleData)
model

}
